Make libreoffice-{bsh,rhino} depend on full java, not just java-headless
Those two packages' "Tools - Macros - Organize Macros -
{BeanShell,JavaScript}..." each have an "Edit" button that wants to bring up a
Swing UI, so silently fails with a Java exception if full java does not happen
to be installed in addition to libreoffice-core's dependency on java-headless.

In theory, a user could always have other LO Java extensions that require non-
headless, so the "correct" approach would be to make libreoffice-core depend on
full java.  But to avoid dependency bloat, only make those sub-packages require
it that obviously need it.

(The ">= 1:1.6" is carried over from libreoffice-core's java-headless
requirement for consistency, even though it is probably cargo-cult by now.)