boost is not part of the Fedora Flatpak SDK after all
...contrary to what eefb0e3720c126ba17b561e0fdfabc05cefac88d "boost is part of
the Fedora, but not of the RHEL Flatpak SDK" had claimed (and which this commit
thus effectively reverts).  In Fedora, boost is part of flatpak-common, but
which is built with an /app, not a /usr prefix.

(And as we no longer append to flatpakoptions anywhere, we can switch back from
%global to %define.)